update tipe data

+++ database/migrations/2024_05_06_100550_create_results_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('diagnosa_id')->constrained('diagnosa')->cascadeOnDelete();
             $table->foreignId('penyakit_id')->constrained('penyakit')->cascadeOnDelete();
-            $table->integer('persentase');
+            $table->double('persentase');
             $table->timestamps();
         });
     }
